Product

+++ product/cmd/domain/model/product_sku.go
@@ -9,7 +9,7 @@ type ProductSku struct {
 	SellPrice           float64 `gorm:"column:sell_price"`
 	CostPrice           float64 `gorm:"column:cost_price;default:1"`
 	Stock               int32   `gorm:"column:stock"`
-	StockWarn           int     `gorm:"column:stock_warn"`
+	StockWarn           int64   `gorm:"column:stock_warn"`
 }
 
 func (p *ProductSku) TableName() string {
